backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.4.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
